Android UI-test refactoring
